20230804mypage객체구현테스트2

+++ src/main/java/com/source/plusutil/mypage/dto/MyPageRequestDto.kt
@@ -1,4 +1,4 @@
 package com.source.plusutil.mypage.dto
 
-data class MyPageRequestDto(val userId : Long) {
+data class MyPageRequestDto(val userId : Int) {
 }
